update names of controllers and models to match naming conventions. add index files to each subdirectory
