+++ src/controllers/index.js
@@ -1,0 +1,6 @@
+/* eslint-disable import/prefer-default-export */
+import requestController from './requestController';
+
+export {
+  requestController,
+};
