+++ src/models/index.js
@@ -1,0 +1,4 @@
+/* eslint-disable import/prefer-default-export */
+import RequestModel from './requestModel';
+
+export { RequestModel as Request };
