+++ src/routers/index.js
@@ -1,0 +1,6 @@
+/* eslint-disable import/prefer-default-export */
+import router from './requestRouter';
+
+export {
+  router,
+};
